requirements.txt
streamlit
pandas
requests
beautifulsoup4
pillow
cloudscraper
pytesseract
